finished solving segfault
